Create a new version that fixes a bug when multiple swagger clients are in use.
